Reimplement git source
Poetry is using dulwich to pull and manage package Git clones.
There are performance issues associated with this (speed of clones and
the size of the cloned repos), but more importantly, dulwich is buggy
and often fails with errors like `AttributeError: 'Tag' object has
no attribute 'parents'` (being confused by annotated tags it seems).

A better way is to revert back to using the actual Git binary for
everything.